// client/src/app/modules/quotation/model/ProvisionBoardDisplayedResult.ts
export interface ProvisionBoardDisplayedResult {
  employee: number;

	nbProvisionNew: number;
	nbProvisionInProgress: number;
	nbProvisionWaiting: number;
	nbProvisionWaitingGreffe: number;
	nbProvisionWaitingAuthority: number;
	nbProvisionConfrerePublished: number;
	nbProvisionValidateGreffe: number;
	nbProvisionRefusedGreffe: number;
	nbProvisionPublished: number;
	nbProvisionDone: number;
}